# backend/client/utente-client/fascicolo/anagrafica/api-doc.json
{
	"swagger": "2.0",
	"info": {
		"description": "Modulo per l'erogazione dei servizi di anagrafica del fascicolo",
		"version": "1.0.0",
		"title": "A4G - Fascicolo - Anagrafica"
	},
	"host": "itts55.intra.infotn.it:18080",
	"basePath": "/anagrafica",
...
	"paths": {
...
		"/api/v1/detenzione/{cuaa}": {
			"get": {
				"tags": [
					"detenzione-controller"
				],
				"summary": "Dato un cuaa restituisce le detenzioni di un fascicolo",
				"operationId": "getDetenzioneUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "cuaa",
						"in": "path",
						"description": "CUAA",
						"required": true,
						"type": "string"
					},
					{
						"name": "idValidazione",
						"in": "query",
						"description": "idValidazione",
						"required": false,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/DetenzioneDto"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
...
		"/api/v1/fascicolo/{cuaa}/check-lettura-fascicolo": {
			"get": {
				"tags": [
					"fascicolo-controller"
				],
				"summary": "Controllo se l'utente autenticato ha i permessi sul fascicolo specificato da cuaa",
				"operationId": "checkLetturaFascicoloUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "cuaa",
						"in": "path",
						"description": "CUAA azienda agricola",
						"required": true,
						"type": "string"
					},
					{
						"name": "idValidazione",
						"in": "query",
						"description": "idValidazione",
						"required": false,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "boolean"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
...
		"/api/v1/fascicolo/{cuaa}/unita-tecnico-economiche": {
			"get": {
				"tags": [
					"fascicolo-controller"
				],
				"summary": "Servizio che permette di reperire le unita tecnico economiche associate al cuaa",
				"operationId": "unitaLocaliUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "cuaa",
						"in": "path",
						"description": "Codice fiscale di cui si vuole reperire le unita tecnico economiche",
						"required": true,
						"type": "string"
					},
					{
						"name": "idValidazione",
						"in": "query",
						"description": "idValidazione",
						"required": false,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/UnitaTecnicoEconomicheDto"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
...
		"/api/v1/personafisica/{codiceFiscale}": {
			"get": {
				"tags": [
					"persona-fisica-controller"
				],
				"summary": "getPersonaFisica",
				"operationId": "getPersonaFisicaUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "codiceFiscale",
						"in": "path",
						"description": "Codice Fiscale Persona",
						"required": true,
						"type": "string"
					},
					{
						"name": "idValidazione",
						"in": "query",
						"description": "idValidazione",
						"required": false,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/PersonaFisicaDto"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
		"/api/v1/personagiuridica/{codiceFiscale}": {
			"get": {
				"tags": [
					"persona-giuridica-controller"
				],
				"summary": "getPersonaFisica",
				"operationId": "getPersonaFisicaUsingGET_1",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "codiceFiscale",
						"in": "path",
						"description": "Codice Fiscale Persona",
						"required": true,
						"type": "string"
					},
					{
						"name": "idValidazione",
						"in": "query",
						"description": "idValidazione",
						"required": false,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/PersonaGiuridicaDto"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
...
	},
	"definitions": {
		"AnagraficaDto": {
			"type": "object",
			"required": [
				"cognome",
				"dataNascita",
				"deceduto",
				"nome",
				"sesso"
			],
			"properties": {
				"cognome": {
					"type": "string",
					"description": "Cognome della persona"
				},
				"comuneNascita": {
					"type": "string",
					"description": "Comune di nascita della persona della persona"
				},
				"dataNascita": {
					"type": "string",
					"format": "date",
					"description": "Data di nascita della persona"
				},
				"deceduto": {
					"type": "boolean",
					"description": "Se si tratta di una persona deceduta"
				},
				"nome": {
					"type": "string",
					"description": "Nome della persona"
				},
				"provinciaNascita": {
					"type": "string",
					"description": "Provincia di nascita della persona"
				},
				"sesso": {
					"type": "string",
					"description": "Sesso della persona"
				}
			},
			"title": "AnagraficaDto"
		},
		"AttivitaAtecoDto": {
			"type": "object",
			"properties": {
				"codice": {
					"type": "string"
				},
				"descrizione": {
					"type": "string"
				},
				"fonte": {
					"type": "string"
				},
				"importanza": {
					"type": "string"
				}
			},
			"title": "AttivitaAtecoDto",
			"description": "Attivita ateco"
		},
		"AttivitaDto": {
			"type": "object",
			"required": [
				"codice",
				"descrizione",
				"fonteDato",
				"importanza"
			],
			"properties": {
				"codice": {
					"type": "string",
					"description": "Codifica ateco"
				},
				"descrizione": {
					"type": "string",
					"description": "Descrizione dell'attivita"
				},
				"fonteDato": {
					"type": "string",
					"description": "Indica il livello di importanza dell'attivita",
					"enum": [
						"ANAGRAFE_TRIBUTARIA",
						"CAMERA_COMMERCIO"
					]
				},
				"importanza": {
					"type": "string",
					"description": "Indica il livello di importanza dell'attivita",
					"enum": [
						"PRIMARIO_IMPRESA",
						"PRIMARIO",
						"SECONDARIO"
					]
				}
			},
			"title": "AttivitaDto"
		},
...
		"DestinazioniUsoDto": {
			"type": "object",
			"properties": {
				"descrizione": {
					"type": "string"
				}
			},
			"title": "DestinazioniUsoDto",
			"description": "Destinazione uso"
		},
		"DetenzioneDto": {
			"type": "object",
			"properties": {
				"codiceFiscale": {
					"type": "string",
					"description": "Codice fiscale"
				},
				"dataFine": {
					"type": "string",
					"format": "date",
					"description": "Dati fine detenzione"
				},
				"dataInizio": {
					"type": "string",
					"format": "date",
					"description": "Data inizio detenzione"
				},
				"denominazione": {
					"type": "string",
					"description": "Denominazione fascicolo"
				},
				"id": {
					"type": "integer",
					"format": "int64",
					"description": "Id detenzione"
				},
				"tipoDetenzione": {
					"type": "string",
					"description": "Tipo di detenzione del fascicolo (e.g. 'mandato' oppure 'detenzione in proprio')",
					"enum": [
						"MANDATO",
						"DETENZIONE_IN_PROPRIO"
					]
				}
			},
			"title": "DetenzioneDto",
			"description": "Detenzione di un fascicolo "
		},
...
		"ImpresaIndividualeDto": {
			"type": "object",
			"properties": {
				"denominazione": {
					"type": "string"
				},
				"formaGiuridica": {
					"type": "string"
				},
				"oggettoSociale": {
					"type": "string"
				},
				"partitaIva": {
					"type": "string"
				},
				"sedeLegale": {
					"$ref": "#/definitions/SedeDto"
				}
			},
			"title": "ImpresaIndividualeDto"
		},
		"IndirizzoDto": {
			"type": "object",
			"properties": {
				"cap": {
					"type": "string"
				},
				"civico": {
					"type": "string"
				},
				"codiceIstat": {
					"type": "string"
				},
				"comune": {
					"type": "string"
				},
				"frazione": {
					"type": "string"
				},
				"localita": {
					"type": "string"
				},
				"provincia": {
					"type": "string"
				},
				"toponimo": {
					"type": "string"
				},
				"via": {
					"type": "string"
				}
			},
			"title": "IndirizzoDto"
		},
...
		"IscrizioneRepertorioEconomicoDto": {
			"type": "object",
			"required": [
				"dataIscrizione"
			],
			"properties": {
				"cessata": {
					"type": "boolean"
				},
				"codiceRea": {
					"type": "integer",
					"format": "int64"
				},
				"dataIscrizione": {
					"type": "string",
					"format": "date",
					"description": "Data di iscrizione"
				},
				"provinciaRea": {
					"type": "string"
				}
			},
			"title": "IscrizioneRepertorioEconomicoDto"
		},
		"IscrizioneSezioneDto": {
			"type": "object",
			"properties": {
				"coltivatoreDiretto": {
					"type": "string"
				},
				"dataIscrizione": {
					"type": "string",
					"format": "date"
				},
				"qualifica": {
					"type": "string"
				},
				"sezione": {
					"type": "string",
					"enum": [
						"ORDINARIA",
						"SPECIALE"
					]
				}
			},
			"title": "IscrizioneSezioneDto"
		},
...
		"PersonaFisicaDto": {
			"type": "object",
			"properties": {
				"anagrafica": {
					"description": "I dati dell'eventuale ditta individuale",
					"$ref": "#/definitions/AnagraficaDto"
				},
				"codiceFiscale": {
					"type": "string"
				},
				"domicilioFiscale": {
					"description": "Domicilio fiscale della persona fisica",
					"$ref": "#/definitions/IndirizzoDto"
				},
				"impresaIndividuale": {
					"description": "I dati dell'eventuale ditta individuale",
					"$ref": "#/definitions/ImpresaIndividualeDto"
				},
				"iscrizioniSezione": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/IscrizioneSezioneDto"
					}
				}
			},
			"title": "PersonaFisicaDto"
		},
...
		"PersonaGiuridicaDto": {
			"type": "object",
			"properties": {
				"capitaleSocialeDeliberato": {
					"type": "number",
					"format": "double"
				},
				"codiceFiscale": {
					"type": "string"
				},
				"dataCostituzione": {
					"type": "string",
					"format": "date"
				},
				"dataTermine": {
					"type": "string",
					"format": "date"
				},
				"denominazione": {
					"type": "string"
				},
				"formaGiuridica": {
					"type": "string"
				},
				"iscrizioniSezione": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/IscrizioneSezioneDto"
					}
				},
				"oggettoSociale": {
					"type": "string"
				},
				"partitaIva": {
					"type": "string"
				},
				"rappresentanteLegale": {
					"description": "Rappresentante legale dell'impresa (valorizzato solo per anagrafe tributaria)",
					"$ref": "#/definitions/RappresentanteLegaleDto"
				},
				"sedeLegale": {
					"$ref": "#/definitions/SedeDto"
				}
			},
			"title": "PersonaGiuridicaDto"
		},
...
		"RappresentanteLegaleDto": {
			"type": "object",
			"required": [
				"codiceFiscale",
				"nominativo"
			],
			"properties": {
				"codiceFiscale": {
					"type": "string",
					"description": "Codice fiscale del rappresentante legale"
				},
				"nominativo": {
					"type": "string",
					"description": "Nominativo del rappresentante legale"
				}
			},
			"title": "RappresentanteLegaleDto",
			"description": "Codice fiscale del rappresentante legale"
		},
...
		"SedeDto": {
			"type": "object",
			"required": [
				"indirizzo"
			],
			"properties": {
				"attivitaAteco": {
					"type": "array",
					"description": "Attivita ateco della ditta",
					"items": {
						"$ref": "#/definitions/AttivitaDto"
					}
				},
				"indirizzo": {
					"description": "Indirizzo della sede",
					"$ref": "#/definitions/IndirizzoDto"
				},
				"indirizzoCameraCommercio": {
					"description": "Indirizzo della sede (Camera Di Commercio)",
					"$ref": "#/definitions/IndirizzoDto"
				},
				"indirizzoPec": {
					"type": "string",
					"description": "Email della sede"
				},
				"iscrizioneRegistroImprese": {
					"description": "Dati di registrazione al registro imprese",
					"$ref": "#/definitions/IscrizioneRepertorioEconomicoDto"
				},
				"telefono": {
					"type": "string",
					"description": "Telefono della sede"
				}
			},
			"title": "SedeDto"
		},
...
		"UnitaTecnicoEconomicheDto": {
			"type": "object",
			"properties": {
				"attivita": {
					"type": "string"
				},
				"attivitaAteco": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/AttivitaAtecoDto"
					}
				},
				"cap": {
					"type": "string"
				},
				"causaleCessazione": {
					"type": "string"
				},
				"codiceIstatComune": {
					"type": "string"
				},
				"comune": {
					"type": "string"
				},
				"dataApertura": {
					"type": "string",
					"format": "date"
				},
				"dataCessazione": {
					"type": "string",
					"format": "date"
				},
				"dataDenunciaCessazione": {
					"type": "string",
					"format": "date"
				},
				"dataDenunciaInizioAttivita": {
					"type": "string",
					"format": "date"
				},
				"destinazioniUso": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/DestinazioniUsoDto"
					}
				},
				"frazione": {
					"type": "string"
				},
				"identificativoUte": {
					"type": "string"
				},
				"indirizzo": {
					"type": "string"
				},
				"indirizzoPec": {
					"type": "string"
				},
				"numeroCivico": {
					"type": "string"
				},
				"provincia": {
					"type": "string"
				},
				"settoreMerceologico": {
					"type": "string"
				},
				"telefono": {
					"type": "string"
				},
				"toponimo": {
					"type": "string"
				},
				"via": {
					"type": "string"
				}
			},
			"title": "UnitaTecnicoEconomicheDto",
			"description": "Unita tecnico economiche"
		},
...
	}
}